Create Introduction
As a junior data analyst working on the marketing analyst team at Cyclistic, a bike-share company in Chicago. The director of the marketing believes the future of the company is maxmizing more memberships, so the goal for the current business is coverting casual riders into annual members. To achieve this goal, the marketing team is going to analyzing the Cyclistic historical bike trip data to identify trends.